🧹 chore: #2 reorganize gemini  test files
