+++ language/gemini.go
@@ -1,6 +1,6 @@
 // language/gemini.go
 
-package language
+package translation
 
 import (
 	"bytes"
@@ -11,6 +11,21 @@ import (
 	"os"
 	"regexp"
 )
+
+
+// request
+type TranslationRequest struct {
+	Text string `json:"text"`
+	// Text       []string `json:"text"
+	TargetLang string `json:"target_lang"` // ex: "ko", "en", "fr"
+}
+
+// response
+type TranslationResponse struct {
+	TranslatedText string `json:"translated_text"`
+	// TranslatedText []string `json:"translated_text"`
+}
+
 
 func TranslateHandler(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodPost {
